split integration, prepare for more statuspage integrations
